Fix build and runtime error
